Extra Credit: Merged Records Into AbstractRecord and Record Files
