+++ Logger.Tests/BaseLoggerMixinsTests.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 
+[assembly: CLSCompliant(true)]
 namespace Logger.Tests;
 
 [TestClass]
