--- Logger/Student.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Logger
-{
-    public record Student(FullName FullName, string Degree, string SchoolName) : Person(FullName)
-    { 
- 
-    }
-  
-}
